Ticket #992: omit xml declaration from XQuery output, if the output result is HTML or TEXT.

+++ src/eionet/gdem/utils/Utils.java
@@ -172,7 +172,8 @@ public class Utils {
 
   public static void deleteFile(String fName) {
     File f = new File(fName);
-    f.delete();
+    if(f.exists())
+    	f.delete();
 
   }
